third_party/blink/web_tests/fast/filesystem/workers/simple-persistent-expected.txt

[Worker] requestFileSystem PERSISTENT test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/simple-persistent.js
[Worker] Successfully obtained PERSISTENT FileSystem:file__0:Persistent
PASS [Worker] fileSystem.name.length > 0 is true
PASS [Worker] fileSystem.root.fullPath is "/"
PASS successfullyParsed is true

TEST COMPLETE
